// frontend/src/api/models/JobOut.ts
export interface JobOut {
    /**
     * 
     * @type {string}
     * @memberof JobOut
     */
    jobId: string;
    /**
     * 
     * @type {JobType}
     * @memberof JobOut
     */
    jobType: JobType;
    /**
     * 
     * @type {string}
     * @memberof JobOut
     */
    promptSlug?: string | null;
    /**
     * 
     * @type {string}
     * @memberof JobOut
     */
    promptDisplayName?: string | null;
    /**
     * 
     * @type {string}
     * @memberof JobOut
     */
    projectId: string;
    /**
     * 
     * @type {JobStatus}
     * @memberof JobOut
     */
    status: JobStatus;
    /**
     * 
     * @type {string}
     * @memberof JobOut
     */
    celeryTaskId?: string | null;
    /**
     * 
     * @type {{ [key: string]: any; }}
     * @memberof JobOut
     */
    result?: { [key: string]: any; } | null;
    /**
     * 
     * @type {string}
     * @memberof JobOut
     */
    triggeredByUserId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof JobOut
     */
    triggeredBy?: string | null;
    /**
     * 
     * @type {string}
     * @memberof JobOut
     */
    createdAt?: string | null;
    /**
     * 
     * @type {string}
     * @memberof JobOut
     */
    updatedAt?: string | null;
}